+++ test/function/matrix/ones.test.js
@@ -55,7 +55,7 @@ describe('ones', function() {
     assert.deepEqual(ones([2,3,4]), res);
   });
 
-  // TODO: test setting matrix.defaultType
+  // TODO: test setting `matrix`
 
   it('should create a matrix with ones with the same size as original matrix', function () {
     var a = matrix([[1, 2, 3], [4, 5, 6]]);